Rename "community challenge"
The schema has CommunityChoice challenge, stick with this naming.
