+++ vit-servicing-station-lib/src/db/models/proposals/community_choice.rs
